+++ src/main/java/h99/ecommerce/controller/OrderController.java
@@ -21,8 +21,8 @@ public class OrderController {
      */
     @PostMapping
     public ResponseEntity<Order> createOrder(
-            @RequestParam Integer userId,
-            @RequestParam(required = false) Integer userCouponId
+            @RequestParam Long userId,
+            @RequestParam(required = false) Long userCouponId
     ) {
         if (userId == null || userId <= 0) {
             return ResponseEntity.badRequest().build();
@@ -39,7 +39,7 @@ public class OrderController {
      * 주문 상세 조회
      */
     @GetMapping("/{orderId}")
-    public ResponseEntity<Order> getOrder(@PathVariable Integer orderId) {
+    public ResponseEntity<Order> getOrder(@PathVariable Long orderId) {
         if (orderId == null || orderId <= 0) {
             return ResponseEntity.badRequest().build();
         }
@@ -52,7 +52,7 @@ public class OrderController {
      * 사용자별 주문 목록 조회
      */
     @GetMapping("/users/{userId}")
-    public ResponseEntity<List<Order>> getOrdersByUserId(@PathVariable Integer userId) {
+    public ResponseEntity<List<Order>> getOrdersByUserId(@PathVariable Long userId) {
         if (userId == null || userId <= 0) {
             return ResponseEntity.badRequest().build();
         }
